pyproject.toml
[project]
name = "morphic"
version = "1.0.0"
description = "Unified media toolkit: format conversion, duplicate detection, EXIF inspection, batch resizing, and file organisation"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Andreas Sagen" }
]
keywords = [
    "file", "image", "video", "converter", "scanner",
    "duplicate", "finder", "batch", "media", "exif",
    "resize", "organizer",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Multimedia :: Video :: Conversion",
    "Topic :: Utilities",
]

dependencies = [
    "flask>=3.0.0",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "ImageHash>=4.3.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "scipy>=1.10.0",
    "piexif>=1.1.3",
]

[project.optional-dependencies]
heif = [
    "pillow-heif>=0.13.0",
]
gpu = [
    "torch>=2.10.0",
    "cupy-cuda12x>=14.0.0",
    "pyopencl>=2023.1",
]
[project.scripts]
morphic = "morphic.frontend.app:main"

[project.urls]
Homepage = "https://github.com/exTerEX/morphic"
Repository = "https://github.com/exTerEX/morphic"
Issues = "https://github.com/exTerEX/morphic/issues"

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

# ── UV configuration ──────────────────────────────────────────────────────────

[tool.uv]
package = true

[tool.uv.build-backend]
module-name = "morphic"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pyright>=1.0.0",
    "sphinx>=7.0.0",
    "furo>=2024.0.0",
    "sphinx-copybutton>=0.5.0",
]

# ── Ruff configuration ───────────────────────────────────────────────────────

[tool.ruff]
respect-gitignore = true
line-length = 79
indent-width = 4

[tool.ruff.lint]
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 79
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.lint.isort]
section-order = [
    "future",
    "standard-library",
    "first-party",
    "local-folder",
    "third-party",
]

# ── Mypy configuration ──────────────────────────────────────────────────────

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

# ── PyTest configuration ────────────────────────────────────────────────────

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = ["-v", "--tb=short", "--strict-markers"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

# ── Coverage configuration ──────────────────────────────────────────────────

[tool.coverage.run]
source = ["src/morphic"]
branch = true
omit = ["*/tests/*", "*/__init__.py", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "def main\\(",
]
show_missing = true

# ── PyRight configuration ───────────────────────────────────────────────────

[tool.pyright]
include = ["src"]
typeCheckingMode = "basic"
